moved publish button to hero component for listing details page
